[chore](check) open shorten-64-to-32 error  (#41197)
```
doris/be/src/pipeline/pipeline_fragment_context.cpp:1160:76: error: implicit conversion loses integer precision: 'size_type' (aka 'unsigned long') to 'int' [-Werror,-Wshorten-64-to-32]
 1159 |         _sink.reset(new MultiCastDataStreamSinkOperatorX(
      |                         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 1160 |                 sink_id, sources, thrift_sink.multi_cast_stream_sink.sinks.size(), pool,
      |                                   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~
```

+++ be/src/vec/core/block.cpp
@@ -854,7 +854,7 @@ Status Block::append_to_block_by_selector(MutableBlock* dst,
 }
 
 Status Block::filter_block(Block* block, const std::vector<uint32_t>& columns_to_filter,
-                           int filter_column_id, int column_to_keep) {
+                           size_t filter_column_id, size_t column_to_keep) {
     const auto& filter_column = block->get_by_position(filter_column_id).column;
     if (const auto* nullable_column = check_and_get_column<ColumnNullable>(*filter_column)) {
         const auto& nested_column = nullable_column->get_nested_column_ptr();
@@ -892,7 +892,7 @@ Status Block::filter_block(Block* block, const std::vector<uint32_t>& columns_to
     return Status::OK();
 }
 
-Status Block::filter_block(Block* block, int filter_column_id, int column_to_keep) {
+Status Block::filter_block(Block* block, size_t filter_column_id, size_t column_to_keep) {
     std::vector<uint32_t> columns_to_filter;
     columns_to_filter.resize(column_to_keep);
     for (uint32_t i = 0; i < column_to_keep; ++i) {
